bleah

+++ DFC.App.SkillsHealthCheck/Models/SessionDataModel.cs
@@ -5,7 +5,7 @@ namespace DFC.App.SkillsHealthCheck.Models
 {
     public class SessionDataModel
     {
-        public long DocumentId { get; set; }
+        public int DocumentId { get; set; }
 
         public Dictionary<string, AssessmentQuestionsOverView> AssessmentQuestionsOverViews { get; set; }
